+++ docs/fr_FR/changelog.md
@@ -4,12 +4,19 @@
 >
 >Pour rappel s'il n'y a pas d'information sur la mise à jour, c'est que celle-ci concerne uniquement de la mise à jour de documentation, de traduction ou de texte.
 
+# 16/11/2022
+Ajout valorisation et affichage des commandes de type info dans l'onglet des commandes
+Gestion envoi de notification en fonction du journal d'évènement
+
+* activation / désactivation
+* intrusion
+* porte ouverte et alarme activée
+
 # 21/10/2022
 Ne plus afficher dans la log le corps du retour de l'api IMA si celui-ci n'est pas de type JSON
 
 # 20/10/2022
 Correction bug récupération des tokens IMA
-
 # 08/11/2021
 - Modification gestion des jetons de session (stockage en BDD)
 - possibilité de supprimer les jetons de session (permet de réinitialiser les appels aux api IMA Protect)